# Logging: detect log level through layers of wrappers