--- a/pkg/util/log/slogadapter.go
+++ b/pkg/util/log/slogadapter.go
@@ -12,20 +12,20 @@ import (
 // SlogFromGoKit returns slog adapter for logger.
 func SlogFromGoKit(logger log.Logger) *slog.Logger {
 	var sl slog.Level
-	x, ok := logger.(leveledLogger)
-	if !ok {
+	x := privateLevelDetector{
+		string:   "This struct is expected to be used with levelLogger only",
+		logLevel: debugLevel,
+	}
+	logger.Log("test", &x)
+	switch x.logLevel {
+	case infoLevel:
+		sl = slog.LevelInfo
+	case warnLevel:
+		sl = slog.LevelWarn
+	case errorLevel:
+		sl = slog.LevelError
+	default:
 		sl = slog.LevelDebug
-	} else {
-		switch x.level() {
-		case infoLevel:
-			sl = slog.LevelInfo
-		case warnLevel:
-			sl = slog.LevelWarn
-		case errorLevel:
-			sl = slog.LevelError
-		default:
-			sl = slog.LevelDebug
-		}
 	}
 
 	lvl := slog.LevelVar{}