--- a/pkg/util/log/slogadapter_test.go
+++ b/pkg/util/log/slogadapter_test.go
@@ -8,6 +8,7 @@ import (
 	"log/slog"
 	"testing"
 
+	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	dslog "github.com/grafana/dskit/log"
 	"github.com/stretchr/testify/assert"
@@ -90,3 +91,29 @@ func TestSlogFromGoKit(t *testing.T) {
 		}
 	})
 }
+
+func BenchmarkSlogFromGoKit(b *testing.B) {
+	// Set up a logger with a stack of wrappers, similar to what tsdb gets.
+	var lvl dslog.Level
+	_ = lvl.Set("info")
+	var logger log.Logger
+	logger = noopLogger{}
+	logger = newFilter(logger, lvl)
+	logger = WithUserID("userID", logger)
+
+	sl := SlogFromGoKit(logger)
+	b.Run("log", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			sl.Info("msg", "foo", "bar", "more", "data")
+		}
+	})
+	b.Run("debuglog", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			sl.Debug("msg", "foo", "bar", "more", "data")
+		}
+	})
+}
+
+type noopLogger struct{}
+
+func (noopLogger) Log(...interface{}) error { return nil }