#### What this PR does

Expose log level through log wrappers, so that `Debug()` calls are fast when debug logging is disabled.

It works by making a `Log()` call so it reaches the bottom layer.

However, if `SlogFromGoKit` is used with a logger outside of this package, you will get a spurious log line.

#### Which issue(s) this PR fixes or relates to

Inspired by https://github.com/prometheus/prometheus/pull/15993

This is a competitor to #10604.  Credit @dimitarvdimitrov for the idea.

Before:
```
BenchmarkSlogFromGoKit/log-28             981555              1176 ns/op            1008 B/op         16 allocs/op
BenchmarkSlogFromGoKit/debuglog-28        949378              1175 ns/op            1008 B/op         16 allocs/op
BenchmarkDebugLog-28                    26430567                44.55 ns/op           96 B/op          2 allocs/op
```

After:
```
BenchmarkSlogFromGoKit/log-28             994729              1165 ns/op            1008 B/op         16 allocs/op
BenchmarkSlogFromGoKit/debuglog-28      358293070                3.342 ns/op           0 B/op          0 allocs/op
BenchmarkDebugLog-28                    24998392                45.67 ns/op           96 B/op          2 allocs/op
```

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.