@dimitarvdimitrov 2025-02-14
for the record: I prefer this approach because the other one requires committers be aware of and keep the conventions in the package (of using `wrap` in #10604)